refactor: :recycle: Move, rename all component page to pages folder and update the imports in router.tsx
