+++ src/pages/Home/TreesHighlights.tsx
